Fix conditional imports if `mkIf` is at top-level.
Technically conditional imports are already implemented,
since you could write modules inline into `imports`.

I.e. code like:

  {
    imports = [
      (lib.mkIf some-condition some-module-config)
    ];
  }

was already allowed. Just not paths instead of `some-module-config`.

Furthermore the new implementation also allows
any other option and just includes them unconditionally.
In practice this means that meta information, option definitions and
disabledModules are always included.
It could be argued that this is unintuitiv, but the extra
cost of implementing this "correctly" (including nested `mkIf`s)
and the problem of effectivly turning of those features
in submodules as well is an even bigger negative,
so this is the solution I went with.

Additionally, since we don't force the config values
into a `config`, having a `lib.mkIf ... { config = ...; }`
also works just fine.